feat: project architecture setup.
